requirements.txt
beautifulsoup4~=4.8.0
lxml~=4.6.5
pre-commit
pytest
pytest-cov
PyYAML~=6.0
requests~=2.22.0
selenium~=4.1.5
